NGS-general: relocate 'sam2soap.py' to bcftbx.cli.
