+++ bin/sam2soap.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     sam2soap.py: convert from SAM to SOAP format
+#     Copyright (C) University of Manchester 2021 Peter Briggs
+#
+from bcftbx.cli.sam2soap import main
+if __name__ == "__main__":
+     main()
